# 8302204: Optimize BigDecimal.divide