@kuksenko 2023-02-11
"The performance looks good." - Could you support this statement with some benchmark results?
Thank you.