@weixlu 2023-02-13
> As for TPC-DS [AUTO-RESULT] QueryTotal=1968s vs [AUTO-RESULT] QueryTotal=1934s that gives ~1.7% of performance difference. Are you sure that this small diff is a real diff, but not run-to-run variance?

These queries are separate and independent. We can even config TPC-DS to run query4 alone and the result is 100s vs 71s. The other queries don't have much to do with decimals. So if you look at the total time, yes there is just a slight diff, because only query4 is impacted by our optimization. But if you just look at query4, the acceleration is shown consistently, we have seen data like 102s->68s, 100s->71s.